first commit RocketShip project
This project showcases features such as assembling multiple Rocket stages to build an integrated RocketShip. It also potrays flight using the staging of the attached Rocket stages and burning these stages to achieve a Thrust Vector and hence a flight. This is made to work for a multiplayer networked environment.